Add BaseMetadata dataclass
