+++ netspresso/metadata/trainer.py
@@ -1,9 +1,8 @@
-import json
-from dataclasses import asdict, dataclass, field
+from dataclasses import dataclass, field
 from typing import Dict, List
 
-from netspresso.enums.metadata import Status, TaskType
-from netspresso.metadata.common import AvailableOption, ExceptionDetail, InputShape
+from netspresso.enums.metadata import TaskType
+from netspresso.metadata.common import AvailableOption, BaseMetadata, InputShape
 
 
 @dataclass
@@ -23,9 +22,7 @@ class TrainingInfo:
 
 
 @dataclass
-class TrainerMetadata:
-    status: Status = Status.IN_PROGRESS
-    message: ExceptionDetail = field(default_factory=ExceptionDetail)
+class TrainerMetadata(BaseMetadata):
     task_type: TaskType = TaskType.TRAIN
     logging_dir: str = ""
     best_fx_model_path: str = ""
@@ -35,19 +32,6 @@ class TrainerMetadata:
     training_info: TrainingInfo = field(default_factory=TrainingInfo)
     traning_result: Dict = field(default_factory=dict)
     available_options: List[AvailableOption] = field(default_factory=list)
-
-    def asdict(self) -> Dict:
-        _dict = json.loads(json.dumps(asdict(self)))
-        return _dict
-
-    def update_status(self, status: Status):
-        self.status = status
-
-    def update_message(self, exception_detail):
-        if isinstance(exception_detail, str):
-            self.message.message = exception_detail
-        else:
-            self.message = ExceptionDetail(**exception_detail)
 
     def update_model_info(self, task, model, dataset, input_shapes):
         self.model_info.task = task
